feat: compile to TensorRT
The distilled student, exported to ONNX and built into a strongly typed engine,
runs four times faster than PyTorch at identical precision and identical
weights: 20.95 ms to 5.28 ms at batch 1, same protocol, same machine. Nothing
about the model changed. The layers stopped being Python dispatches.

That overturns the previous commit's conclusion. It found pruning worth 0% of
latency and explained it by the model being activation-bound. Under TensorRT the
same two checkpoints are 5.28 ms and 4.93 ms -- 6.6%, and a 26% smaller engine.
The dominant cost was neither weights nor activations but framework overhead,
large enough to hide the effect completely. A compression measurement taken in
eager PyTorch was measuring the interpreter.

So compression and deployment become one milestone with the runtime first. A
compression number is a statement about a runtime, and taking it on a different
one measures that other runtime instead.

TensorRT 11 does not build engines the way the plan assumed. That plan was
written against the workflow every tutorial still describes: BuilderFlag.FP16,
BuilderFlag.INT8, an IInt8EntropyCalibrator2 fed representative data. None of it
exists in 11.2. Precision is a property of the graph now, and a strongly typed
network takes each layer's type from the ONNX it was handed. The calibrator
version was written first and thrown away; a test asserts the API that exists,
so a release restoring the flags will fail it rather than leave the explanation
quietly wrong.

FP16 needed the model fixed rather than the builder. exact_arithmetic keeps
three pieces of coordinate arithmetic in fp32 and nothing cast back at their
edges -- invisible under autocast, which re-casts an op at a time, but .half()
leaves the fp32 island to propagate into fp16 attention and fp16 layers. The
guard also upcast its data and not the cell coordinates it multiplies against,
so a half model would have put exactly the lattice under the answer that the
guard exists to prevent. Four boundary casts, and fp32 output is unchanged bit
for bit.

fp16 rather than bf16, measured rather than assumed. Both are sixteen bits;
bf16 spends seven on the mantissa against fp16's ten, which resolves a 40 m map
coordinate to 12.5 cm rather than 1.56 cm. Translation does not notice. Heading
does -- 12.5 cm across a 40 m baseline subtends 0.18 deg, the size of the error
being measured -- and the covariance notices most, its median NEES/dof going
from 0.79 to 1.48. ARCHITECTURE.md now carries the map of which precision runs
where, because nothing did and the question kept coming up.

INT8 is still missing and recorded as missing: it needs a graph carrying Q/DQ
nodes, which is nvidia-modelopt's job.

Also excludes build/ from the remote sync, which was deleting the engines
between the command that built them and the command that timed them.

+++ mapposeformer/model/model.py
@@ -297,9 +297,9 @@ class MapPoseFormer(nn.Module):
             # matching gets easier; letting the gradient run back through a
             # chain of warps would make each pass responsible for the ones
             # after it, which is a much longer path for no extra signal.
-            moved = G.transform_points(delta.detach(), det_xy).view_as(
-                det["pts"]
-            )
+            moved = G.transform_points(
+                delta.detach().to(det_xy.dtype), det_xy
+            ).view_as(det["pts"])
             assign, scores, tokens, pad = self._trunk(det, batch, moved)
             # Solved on the *original* coordinates, so this is the total
             # correction and not an increment -- no composition, and the next
